% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicators}
\alias{indicators}
\title{Metadados dos indicadores RIPSA}
\format{
Uma tibble com as colunas:
\describe{
\item{id}{Código RIPSA do indicador.}
\item{theme}{Tema do indicador.}
\item{dimension}{Dimensão do indicador.}
\item{name}{Nome do indicador.}
\item{ckan_package}{Identificador do pacote CKAN.}
\item{ckan_resource}{Identificador do recurso CKAN.}
\item{ckan_resource_url}{URL da página do recurso CKAN.}
\item{download_url}{URL atual para baixar o arquivo CSV compactado do recurso CKAN.}
\item{spatial_granularity}{Níveis geográficos disponíveis no arquivo de origem, separados por vírgula.}
\item{time_granularity}{Granularidade temporal inferida a partir do arquivo de origem.}
\item{time_start}{Primeira data disponível no arquivo de origem.}
\item{time_end}{Última data disponível no arquivo de origem.}
\item{available_categories}{Códigos padronizados das categorias disponíveis, separados por vírgula.}
\item{available_category_names}{Nomes padronizados das categorias disponíveis, separados por barra vertical.}
}
}
\usage{
indicators
}
\description{
Tabela com os indicadores RIPSA atualmente registrados no pacote.
}
\keyword{datasets}
