% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator.R
\name{list_indicators}
\alias{list_indicators}
\title{Liste os indicadores RIPSA disponíveis}
\usage{
list_indicators()
}
\value{
Uma tibble com uma linha por indicador registrado nos metadados do
pacote.
}
\description{
Liste os indicadores RIPSA disponíveis
}
